parallelize Topology.trim
`Topology.trim` can be quite slow if the levelset is expensive and maxrefine is large.
Although the work is embarrassingly parallel, this was not exploited,
because we can't make a shared memory array of element.Reference objects. With
the newly added `parallel.Manager` we have an alternative way of sharing data
between processes: using pickle and pipes. This patch adds support for parallel
trimming using the `parallel.Manager.list`.