add plabel, vlabel arguments to export.triplot
This patch adds the plabel and vlabel arguments to export.triplot for
the coordinates and values, respectively. For 1D plots the labels are
used for the x and y axes; for 2D plots the plabel is used for both x
and y and the vlabel is used for the colorbar.